After Oversampling
An imbalanced training data set will
certainly affect prediction accuracy. The term "class imbalance"
refers to a significant disparity in class sizes. In this
circumstance, any classifier becomes biased in favor of the
dominant class, ignoring the minority class, resulting in
incorrect predictions. To overcome this limitation, we made all
minority classes equal to the majority. When the data amount is
insufficient, oversampling is a strategy for resampling the data
to address imbalance dataset difficulties. By creating additional
samples, it enlarges the minority class dataset. As a result, we
used Random over Sampling on our datase